Build constants in numpy instead of TensorFlow
Seems to be faster in general.

And we're no longer limited by the 2GB TensorFlow limit
(in eager mode).

+++ nengo_dl/neuron_builders.py
@@ -448,7 +448,9 @@ class StochasticSpikingBuilder(TFNeuronBuilder):
         n_spikes = tf.floor(x)
         frac = x - n_spikes
 
-        n_spikes += tf.cast(tf.random.uniform(frac.shape) < frac, n_spikes.dtype)
+        n_spikes += tf.cast(
+            tf.random.uniform(frac.shape, dtype=frac.dtype) < frac, n_spikes.dtype
+        )
 
         n_spikes *= self.alpha * tf.math.sign(J)
